src/visualsearchservice/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pillow==11.0.0
numpy==2.1.3
tensorflow==2.18.0
faiss-cpu==1.9.0
requests==2.32.3
python-jose[cryptography]==3.3.0
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
